Addressing Sam's code review comments
