+++ mocca-client/src/test/java/com/paypal/mocca/client/MoccaSerializerTest.java
@@ -34,7 +34,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable("bar", String.class, newVar("bar"))
         );
 
-        requestTest(variables, SampleResponseDTO.class,"getOneSample", MoccaUtils.OperationType.Query,
+        requestTest(variables, SampleResponseDTO.class,"getOneSample", OperationType.Query,
                 null, "{\n  \"query\" : \"query{getOneSample(foo: \\\"foo\\\", bar: \\\"bar\\\") {bar foo}}\"\n}");
     }
 
@@ -45,7 +45,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable(sampleRequestDTO, SampleRequestDTO.class, newVar("sampleRequest"))
         );
 
-        requestTest(variables, SampleResponseDTO.class,"getOneSample", MoccaUtils.OperationType.Query,
+        requestTest(variables, SampleResponseDTO.class,"getOneSample", OperationType.Query,
                 null, "{\n  \"query\" : \"query{getOneSample(sampleRequest: {bar: \\\"bar\\\", foo: \\\"foo\\\"}) {bar foo}}\"\n}");
     }
 
@@ -55,7 +55,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable("foo", String.class, newVar("foo"))
         );
 
-        requestTest(variables, SampleResponseDTO.class,"getOneSample", MoccaUtils.OperationType.Mutation,
+        requestTest(variables, SampleResponseDTO.class,"getOneSample", OperationType.Mutation,
                 null, "{\n  \"query\" : \"mutation{getOneSample(foo: \\\"foo\\\") {bar foo}}\"\n}");
     }
 
@@ -67,7 +67,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable(sampleRequestDTO, SampleRequestDTO.class, variable)
         );
 
-        requestTest(variables, SampleResponseDTO.class,"getOneSample", MoccaUtils.OperationType.Query,
+        requestTest(variables, SampleResponseDTO.class,"getOneSample", OperationType.Query,
                 null, "{\n  \"query\" : \"query{getOneSample(sampleRequest: {bar: \\\"bar\\\"}) {bar foo}}\"\n}");
     }
 
@@ -79,7 +79,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable(complexSampleType, ComplexSampleType.class, newVar("complexSample"))
         );
 
-        requestTest(variables, SampleResponseDTO.class, "getOneSample", MoccaUtils.OperationType.Mutation,
+        requestTest(variables, SampleResponseDTO.class, "getOneSample", OperationType.Mutation,
                 null, "{\n  \"query\" : \"mutation{getOneSample(complexSample: {booleanVar: true, complexField: {innerBooleanVar: false, innerIntVar: 77, innerStringVar: \\\"sevenseven\\\"}, intVar: 7, stringVar: \\\"seven\\\"}) {bar foo}}\"\n}");
     }
 
@@ -91,7 +91,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable(complexSampleType, ComplexSampleType.class, newVar("complexSample"))
         );
 
-        requestTest(variables, SampleResponseDTO.class, "getOneSample", MoccaUtils.OperationType.Query,
+        requestTest(variables, SampleResponseDTO.class, "getOneSample", OperationType.Query,
                 null, "{\n  \"query\" : \"query{getOneSample(complexSample: {booleanVar: true, complexField: {innerBooleanVar: false, innerIntVar: 77}, intVar: 7}) {bar foo}}\"\n}");
     }
 
@@ -104,7 +104,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable(complexSampleType, ComplexSampleType.class, variable)
         );
 
-        requestTest(variables, SampleResponseDTO.class, "getOneSample", MoccaUtils.OperationType.Mutation,
+        requestTest(variables, SampleResponseDTO.class, "getOneSample", OperationType.Mutation,
                 null, "{\n  \"query\" : \"mutation{getOneSample(complexSample: {booleanVar: true, complexField: {innerIntVar: 77}, stringVar: \\\"seven\\\"}) {bar foo}}\"\n}");
     }
 
@@ -115,7 +115,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable(sampleRequestDTO, SampleRequestDTO.class, newVar("sampleRequest"))
         );
 
-        requestTest(variables, ComplexSampleType.class,"getABeer", MoccaUtils.OperationType.Query,
+        requestTest(variables, ComplexSampleType.class,"getABeer", OperationType.Query,
                 null, "{\n  \"query\" : \"query{getABeer(sampleRequest: {bar: \\\"bar\\\", foo: \\\"foo\\\"}) {booleanVar complexField {innerBooleanVar innerIntVar innerStringVar} intVar stringVar}}\"\n}");
     }
 
@@ -127,7 +127,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable(sampleRequestDTO, SampleRequestDTO.class, newVar("sampleRequest"))
         );
 
-        requestTest(variables, ComplexSampleType.class,"getABeer", MoccaUtils.OperationType.Query,
+        requestTest(variables, ComplexSampleType.class,"getABeer", OperationType.Query,
                 selectionSet, "{\n  \"query\" : \"query{getABeer(sampleRequest: {bar: \\\"bar\\\", foo: \\\"foo\\\"}) {booleanVar complexField {innerBooleanVar innerIntVar innerStringVar} intVar stringVar}}\"\n}");
     }
 
@@ -139,7 +139,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable(sampleRequestDTO, SampleRequestDTO.class, newVar("sampleRequest"))
         );
 
-        requestTest(variables, ComplexSampleType.class,"getABeer", MoccaUtils.OperationType.Query,
+        requestTest(variables, ComplexSampleType.class,"getABeer", OperationType.Query,
                 selectionSet, "{\n  \"query\" : \"query{getABeer(sampleRequest: {bar: \\\"bar\\\", foo: \\\"foo\\\"}) {booleanVar complexField stringVar}}\"\n}");
     }
 
@@ -155,7 +155,7 @@ public class MoccaSerializerTest {
         List<MoccaSerializer.Variable> variables = Collections.singletonList(
                 new MoccaSerializer.Variable( superComplexSampleType, SuperComplexSampleType.class, newVar("sampleRequest")));
 
-        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", MoccaUtils.OperationType.Query,
+        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", OperationType.Query,
                 null,  "{\n  \"query\" : \"query{getOneComplexSample(sampleRequest: {booleanVar: true, complexField: {innerBooleanVar: false, innerIntVar: 77, innerStringListVar: [\\\"cat\\\", \\\"dog\\\", \\\"monkey\\\"], innerStringVar: \\\"sevenseven\\\"}, complexListVar: [], intVar: 7, stringVar: \\\"seven\\\"}) {bar foo}}\"\n}");
     }
 
@@ -165,7 +165,7 @@ public class MoccaSerializerTest {
                 null, null, Arrays.asList(new String[] {"cat", "dog", "monkey"}));
         List<MoccaSerializer.Variable> variables = Collections.singletonList(
                 new MoccaSerializer.Variable( superComplexSampleType, SuperComplexSampleType.class, newVar("sampleRequest")));
-        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", MoccaUtils.OperationType.Query,
+        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", OperationType.Query,
                 null,  "{\n  \"query\" : \"query{getOneComplexSample(sampleRequest: {booleanVar: true, intVar: 7, stringListVar: [\\\"cat\\\", \\\"dog\\\", \\\"monkey\\\"], stringVar: \\\"seven\\\"}) {bar foo}}\"\n}");
     }
 
@@ -182,7 +182,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable( superComplexSampleType, SuperComplexSampleType.class,
                         newVar("sampleRequest", "complexField.innerStringListVar")));
 
-        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", MoccaUtils.OperationType.Query,
+        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", OperationType.Query,
                 null,  "{\n  \"query\" : \"query{getOneComplexSample(sampleRequest: {booleanVar: true, complexField: {innerBooleanVar: false, innerIntVar: 77, innerStringVar: \\\"sevenseven\\\"}, complexListVar: [], intVar: 7, stringVar: \\\"seven\\\"}) {bar foo}}\"\n}");
     }
 
@@ -204,7 +204,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable( superComplexSampleType, SuperComplexSampleType.class,
                         newVar("sampleRequest", "complexListVar.innerStringListVar")));
 
-        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", MoccaUtils.OperationType.Query,
+        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", OperationType.Query,
                 null,  "{\n  \"query\" : \"query{getOneComplexSample(sampleRequest: {booleanVar: true, complexListVar: [{innerBooleanVar: false, innerIntVar: 77, innerStringVar: \\\"sevenseven\\\"}, {innerBooleanVar: false, innerIntVar: 99, innerStringVar: \\\"numbernine\\\"}], intVar: 7, stringVar: \\\"seven\\\"}) {bar foo}}\"\n}");
     }
 
@@ -224,7 +224,7 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable( superComplexSampleType, SuperComplexSampleType.class,
                         newVar("sampleRequest", "complexField.innerComplexVar.innerBooleanVar")));
 
-        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", MoccaUtils.OperationType.Query,
+        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", OperationType.Query,
                 null,  "{\n  \"query\" : \"query{getOneComplexSample(sampleRequest: {booleanVar: true, complexField: {innerBooleanVar: false, innerComplexVar: {innerIntVar: 99, innerStringListVar: [\\\"bat\\\", \\\"frog\\\", \\\"money\\\"], innerStringVar: \\\"numbernine\\\"}, innerIntVar: 77, innerStringListVar: [\\\"cat\\\", \\\"dog\\\", \\\"monkey\\\"], innerStringVar: \\\"sevenseven\\\"}, intVar: 7, stringVar: \\\"seven\\\"}) {bar foo}}\"\n}");
     }
 
@@ -252,11 +252,11 @@ public class MoccaSerializerTest {
                 new MoccaSerializer.Variable( superComplexSampleType, SuperComplexSampleType.class,
                         newVar("sampleRequest", "complexField.innerComplexListVar.innerStringVar")));
 
-        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", MoccaUtils.OperationType.Query,
+        requestTest(variables, SampleResponseDTO.class, "getOneComplexSample", OperationType.Query,
                 null,  "{\n  \"query\" : \"query{getOneComplexSample(sampleRequest: {booleanVar: true, complexField: {innerBooleanVar: false, innerComplexListVar: [{innerBooleanVar: false, innerIntVar: 99, innerStringListVar: [\\\"bat\\\", \\\"frog\\\", \\\"money\\\"]}, {innerBooleanVar: true, innerIntVar: 666, innerStringListVar: [\\\"rat\\\", \\\"warthog\\\", \\\"nothing\\\"]}], innerIntVar: 77, innerStringListVar: [\\\"cat\\\", \\\"dog\\\", \\\"monkey\\\"], innerStringVar: \\\"sevenseven\\\"}, intVar: 7, stringVar: \\\"seven\\\"}) {bar foo}}\"\n}");
     }
 
-    private void requestTest(List<MoccaSerializer.Variable> variables, Type responseType, String operationName, MoccaUtils.OperationType operationType, SelectionSet selectionSet, String expectedRequest) throws IOException {
+    private void requestTest(List<MoccaSerializer.Variable> variables, Type responseType, String operationName, OperationType operationType, SelectionSet selectionSet, String expectedRequest) throws IOException {
         byte[] requestBytes = moccaSerializer.serialize(variables, responseType, operationName, operationType, selectionSet);
         String actualRequest = new String(requestBytes);
         assertEquals(actualRequest, expectedRequest);
